@@ -14,13 +14,13 @@
   },
   "alias": {
     "react": "../node_modules/react",
-    "react-dom": "../node_modules/react-dom/profiling",
+    "react-dom": "../node_modules/react-dom",
     "scheduler/tracing": "../node_modules/scheduler/tracing-profiling"
   },
   "devDependencies": {
-    "@types/react": "16.14.1",
-    "@types/react-dom": "16.9.10",
-    "parcel": "1.12.4",
-    "typescript": "4.1.2"
+    "@types/react": "18.0.18",
+    "@types/react-dom": "18.0.6",
+    "parcel": "2.7.0",
+    "typescript": "4.8.2"
   }
 }
