@@ -28,7 +28,7 @@
     "test": "cross-env DEBUG_PRINT_LIMIT=50 && tsdx test --passWithNoTests",
     "lint": "tsdx lint src stories",
     "prepare": "tsdx build",
-    "storybook": "start-storybook -p 6006",
+    "storybook": "cross-env NODE_OPTIONS=--openssl-legacy-provider start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "example:start": "cd ./example && yarn start"
   },
@@ -38,7 +38,7 @@
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
       "\\.(css|less)$": "identity-obj-proxy"
     },
-    "testEnvironment": "jest-environment-jsdom-fifteen"
+    "testEnvironment": "jest-environment-jsdom-sixteen"
   },
   "peerDependencies": {
     "antd": "^4.x",
@@ -52,39 +52,46 @@
   },
   "module": "dist/ant-table-extensions.esm.js",
   "devDependencies": {
-    "@ant-design/icons": "4.3.0",
-    "@babel/core": "7.11.1",
-    "@storybook/addon-actions": "6.1.2",
-    "@storybook/addon-docs": "5.3.21",
-    "@storybook/addon-essentials": "6.1.2",
-    "@storybook/addon-info": "5.3.21",
-    "@storybook/addon-links": "6.1.2",
-    "@storybook/addons": "5.3.21",
-    "@storybook/react": "6.1.2",
-    "@testing-library/jest-dom": "5.11.6",
-    "@testing-library/react": "10.4.9",
-    "@testing-library/user-event": "12.2.2",
-    "@types/lodash": "4.14.165",
-    "@types/react": "16.14.1",
-    "@types/react-dom": "16.9.10",
-    "antd": "4.8.4",
-    "babel-loader": "8.1.0",
-    "cross-env": "7.0.2",
+    "@ant-design/icons": "4.7.0",
+    "@babel/core": "7.19.0",
+    "@mdx-js/react": "^2.1.3",
+    "@storybook/addon-actions": "^6.5.10",
+    "@storybook/addon-essentials": "^6.5.10",
+    "@storybook/addon-interactions": "^6.5.10",
+    "@storybook/addon-links": "^6.5.10",
+    "@storybook/addon-postcss": "^2.0.0",
+    "@storybook/builder-webpack4": "^6.5.10",
+    "@storybook/manager-webpack4": "^6.5.10",
+    "@storybook/react": "^6.5.10",
+    "@storybook/testing-library": "^0.0.13",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/lodash": "4.14.184",
+    "@types/papaparse": "^5.3.5",
+    "@types/react": "18.0.18",
+    "@types/react-dom": "18.0.6",
+    "antd": "^4.23.0",
+    "autoprefixer": "^10.4.8",
+    "babel-loader": "8.2.5",
+    "cross-env": "7.0.3",
+    "html-webpack-plugin": "^5.5.0",
     "identity-obj-proxy": "3.0.0",
-    "jest-environment-jsdom-fifteen": "1.0.2",
-    "react": "16.14.0",
+    "jest-environment-jsdom-sixteen": "^2.0.0",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "react": "18.2.0",
     "react-docgen-typescript-loader": "3.7.2",
-    "react-dom": "16.14.0",
-    "react-is": "16.13.1",
-    "ts-loader": "8.0.11",
+    "react-dom": "18.2.0",
+    "react-is": "18.2.0",
+    "ts-loader": "9.3.1",
     "tsdx": "0.14.1",
-    "tslib": "2.0.3",
-    "typescript": "4.1.2"
+    "tslib": "2.4.0",
+    "typescript": "4.8.2"
   },
   "dependencies": {
-    "fuse.js": "^6.4.1",
-    "lodash": "^4.17.20",
-    "lodash-es": "^4.17.15",
-    "papaparse": "^5.2.0"
+    "fuse.js": "^6.6.2",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "papaparse": "^5.3.2"
   }
 }
